# backend/api-docs/swagger.yaml
basePath: /
definitions:
...
  config.Config_MediaServer:
    properties:
      api_token:
        description: Authentication token for accessing the media server.
        type: string
      enable_plex_event_listener:
        description: Whether to enable the Plex event listener for reapplying images
          on refresh. Plex exclusive feature.
        type: boolean
      enable_sort_by_episode_added_date:
        description: Whether to check episodes for added date when getting Media Items.
          This is only for Plex servers.
        type: boolean
      libraries:
        description: List of media server libraries to manage.
        items:
          $ref: '#/definitions/models.LibrarySection'
        type: array
      type:
        description: Type of media server (e.g., plex, emby, jellyfin).
        type: string
      url:
        description: Base URL of the media server. This is either the IP:Port or the
          domain name (e.g., plex.domain.com).
        type: string
      user_id:
        description: User ID for accessing the media server. This is used for Emby
          and Jellyfin servers.
        type: string
    type: object
...
  httpx.JSONResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/logging.LogErrorInfo'
      status:
        example: success
        type: string
    type: object
  httpx.UnauthorizedResponse:
    properties:
      message:
        example: Invalid or expired token
        type: string
    type: object
...
  logging.LogErrorInfo:
    properties:
      detail:
        type: object
      function:
        example: Reload
        type: string
      help:
        example: Check server logs for more details
        type: string
      line_number:
        example: 42
        type: integer
      message:
        example: Invalid or expired token
        type: string
    type: object
...
  models.CollectionItem:
    properties:
      child_count:
        type: integer
      index:
        description: Unique identifier for the collection in Plex
        type: string
      library_title:
        type: string
      media_items:
        items:
          $ref: '#/definitions/models.MediaItem'
        type: array
      rating_key:
        type: string
      summary:
        type: string
      title:
        type: string
      tmdb_id:
        type: string
    type: object
...
  models.DBSavedSet:
    properties:
      id:
        type: string
      selected_types:
        $ref: '#/definitions/models.SelectedTypes'
      user_created:
        type: string
    type: object
  models.ImageFile:
    properties:
      blurhash:
        description: Blurhash string for preview
        type: string
      episode_number:
        description: Present for Titlecards
        type: integer
      file_size:
        description: Size of the file in bytes
        type: integer
      id:
        description: Unique asset ID for the poster file
        type: string
      item_tmdb_id:
        description: TMDB ID of the parent item
        type: string
      language:
        description: Language Name
        type: string
      modified:
        description: Last modified date
        type: string
      season_number:
        description: Present for Season Posters and Titlecards
        type: integer
      src:
        description: Source URL (ID+Date - calculated by MediUX API)
        type: string
      title:
        description: Present for Titlecards
        type: string
      type:
        description: (poster, backdrop, seasonPoster, specialSeasonPoster, titlecard)
        type: string
    type: object
...
  models.LibrarySection:
    properties:
      id:
        description: Unique identifier for the library section.
        type: string
      media_items:
        items:
          $ref: '#/definitions/models.MediaItem'
        type: array
      path:
        description: Path of the library section on the media server.
        type: string
      title:
        description: Title of the library section.
        type: string
      total_size:
        type: integer
      type:
        description: '"movie" or "show"'
        type: string
    type: object
  models.MediaItem:
    properties:
      added_at:
        description: Added timestamp
        type: integer
      content_rating:
        description: Used in MediaItem Details Page - For more information about the
          item
        type: string
      db_saved_sets:
        description: Used in MediaItem Details Page - For poster sets
        items:
          $ref: '#/definitions/models.DBSavedSet'
        type: array
      guids:
        description: Used in MediaItem Details Page - For ratings
        items:
          $ref: '#/definitions/models.MediaItemGuid'
        type: array
      has_mediux_sets:
        description: Used in Home Page - sorting and filtering
        type: boolean
      ignored_in_db:
        description: Whether the item is marked as ignored
        type: boolean
      ignored_mode:
        description: Mode of ignoring (e.g., "always", "temp")
        type: string
      latest_episode_added_at:
        type: integer
      library_title:
        description: Title of the library/section the item belongs to
        type: string
      movie:
        allOf:
        - $ref: '#/definitions/models.MediaItemMovie'
        description: Present if Type is "movie"; Contains file info
      rating_key:
        description: RatingKey is the internal ID from the media server
        type: string
      released_at:
        description: Release date timestamp
        type: integer
      series:
        allOf:
        - $ref: '#/definitions/models.MediaItemSeries'
        description: Present if Type is "show"; Contains seasons and episodes info
      summary:
        description: Summary or description of the media item
        type: string
      title:
        description: Title of the media item
        type: string
      tmdb_id:
        description: TMDB ID of the media item
        type: string
      type:
        description: '"movie" or "show"'
        type: string
      updated_at:
        description: Last updated timestamp
        type: integer
      year:
        description: Release year of the media item
        type: integer
    type: object
  models.MediaItemEpisode:
    properties:
      added_at:
        type: integer
      episode_number:
        type: integer
      file:
        $ref: '#/definitions/models.MediaItemFile'
      rating_key:
        type: string
      season_number:
        type: integer
      title:
        type: string
    type: object
  models.MediaItemFile:
    properties:
      duration:
        type: integer
      path:
        type: string
      size:
        type: integer
    type: object
  models.MediaItemGuid:
    properties:
      id:
        type: string
      provider:
        type: string
      rating:
        type: string
    type: object
  models.MediaItemMovie:
    properties:
      file:
        $ref: '#/definitions/models.MediaItemFile'
    type: object
  models.MediaItemSeason:
    properties:
      episodes:
        items:
          $ref: '#/definitions/models.MediaItemEpisode'
        type: array
      rating_key:
        type: string
      season_number:
        type: integer
      title:
        type: string
    type: object
  models.MediaItemSeries:
    properties:
      episode_count:
        type: integer
      location:
        type: string
      season_count:
        type: integer
      seasons:
        items:
          $ref: '#/definitions/models.MediaItemSeason'
        type: array
    type: object
  models.MediuxUserInfo:
    properties:
      avatar:
        type: string
      follow:
        type: boolean
      hide:
        type: boolean
      id:
        type: string
      total_sets:
        type: integer
      username:
        type: string
    type: object
...
  models.SelectedTypes:
    properties:
      backdrop:
        type: boolean
      poster:
        type: boolean
      season_poster:
        type: boolean
      special_season_poster:
        type: boolean
      titlecard:
        type: boolean
    type: object
  models.SetRef:
    properties:
      date_created:
        description: Creation date
        type: string
      date_updated:
        description: Last updated date
        type: string
      id:
        description: Set ID
        type: string
      images:
        items:
          $ref: '#/definitions/models.ImageFile'
        type: array
      item_ids:
        description: IDs of items in the set
        items:
          type: string
        type: array
      popularity:
        description: Popularity score of the set
        type: integer
      popularity_global:
        description: Global popularity score of the set
        type: integer
      title:
        description: Set Title
        type: string
      type:
        description: Set Type (movie, show, collection, boxset)
        type: string
      user_created:
        description: User who created the set
        type: string
    type: object
  plex.PlexServerConnections:
    properties:
      address:
        type: string
      ipv6:
        type: boolean
      local:
        type: boolean
      port:
        type: integer
      protocol:
        type: string
      relay:
        type: boolean
      uri:
        type: string
    type: object
  plex.PlexServersResponse:
    properties:
      connections:
        items:
          $ref: '#/definitions/plex.PlexServerConnections'
        type: array
      name:
        type: string
      owned:
        type: boolean
    type: object
...
  routes_ms.GetAllCollectionChildrenItems_Response:
    properties:
      collection_item:
        $ref: '#/definitions/models.CollectionItem'
      sets:
        items:
          $ref: '#/definitions/models.SetRef'
        type: array
      user_follow_hide:
        items:
          $ref: '#/definitions/models.MediuxUserInfo'
        type: array
    type: object
...
  routes_plex.CheckAuthStatusWithPlex_Response:
    properties:
      auth_token:
        type: string
      authenticated:
        type: boolean
      connections_available:
        items:
          $ref: '#/definitions/plex.PlexServersResponse'
        type: array
    type: object
  routes_plex.GetPlexPinAndID_Response:
    properties:
      plex_id:
        type: integer
      plex_pin:
        type: string
    type: object
...
  routes_validation.ValidateMediaServerInfo_Request:
    properties:
      media_server:
        $ref: '#/definitions/config.Config_MediaServer'
    type: object
  routes_validation.ValidateMediaServerInfo_Response:
    properties:
      media_server:
        $ref: '#/definitions/config.Config_MediaServer'
      message:
        type: string
      valid:
        type: boolean
    type: object
...
info:
  contact: {}
  title: Aura API
  version: "1.0"
paths:
...
  /api/mediaserver/collections/item:
    get:
      consumes:
      - application/json
      description: Retrieve all child items of a movie collection from the media server,
        along with their associated posters. This endpoint accepts a query parameter
        to identify the collection and returns the child items contained within that
        collection, as well as any relevant poster sets for those items. This allows
        clients to display the contents of a movie collection along with visual representations
        (posters) for each item.
      parameters:
      - description: Rating Key of the Collection
        in: query
        name: rating_key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/routes_ms.GetAllCollectionChildrenItems_Response'
              type: object
        "401":
          description: Unauthorized (only when Auth.Enabled=true)
          schema:
            $ref: '#/definitions/httpx.UnauthorizedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.JSONResponse'
      security:
      - BearerAuth: []
      summary: Get All Movie Collection Children Items And Posters
      tags:
      - MediaServer
...
  /api/oauth/plex:
    get:
      consumes:
      - application/json
      description: Retrieve a new Plex Pin and the associated Plex ID for authentication.
        This endpoint is used to initiate the Plex authentication process by providing
        the necessary credentials for the user to authenticate their Plex account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/routes_plex.GetPlexPinAndID_Response'
              type: object
        "401":
          description: Unauthorized (only when Auth.Enabled=true)
          schema:
            $ref: '#/definitions/httpx.UnauthorizedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.JSONResponse'
      security:
      - BearerAuth: []
      summary: Get Plex Pin and ID
      tags:
      - Plex
    post:
      consumes:
      - application/json
      description: Check if the provided Plex ID has authenticated with Plex and retrieve
        available server connections if authenticated. This endpoint is used during
        the Plex authentication process to verify the user's Plex account and gather
        necessary information for integration.
      parameters:
      - description: Plex ID to check for authentication
        in: query
        name: plex_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/routes_plex.CheckAuthStatusWithPlex_Response'
              type: object
        "401":
          description: Unauthorized (only when Auth.Enabled=true)
          schema:
            $ref: '#/definitions/httpx.UnauthorizedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.JSONResponse'
      security:
      - BearerAuth: []
      summary: Check Plex Pin for Authentication
      tags:
      - Plex
...
  /api/validate/mediaserver:
    post:
      consumes:
      - application/json
      description: Validate the provided media server information by attempting to
        connect to the media server. This endpoint is used during the onboarding process
        to ensure that the media server settings entered by the user are correct and
        that a connection can be established. The response will indicate whether the
        connection was successful and provide details about the media server if it
        was validated successfully.
      parameters:
      - description: Media Server Information to Validate
        in: body
        name: media_server
        required: true
        schema:
          $ref: '#/definitions/routes_validation.ValidateMediaServerInfo_Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/routes_validation.ValidateMediaServerInfo_Response'
              type: object
        "401":
          description: Unauthorized (only when Auth.Enabled=true)
          schema:
            $ref: '#/definitions/httpx.UnauthorizedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.JSONResponse'
      security:
      - BearerAuth: []
      summary: Validate Media Server Info
      tags:
      - Validation
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"